Purposely allow one tslint error
If there are no errors, tslint outputs an empty file instead of a valid
XML file with no entries:
  https://github.com/palantir/tslint/pull/1463
The empty file causes Jenkins to fail. This is fixed in tslint 4.x, but
that requires upgrading some other tools and updating the lint file.
This is a quick fix while we work on other things.