Initial support for showing different data intervals in the focus chart
Split the logic to filter out the data to show in the focus chart to the component.
This allows the main chart to show data of a different interval than the focus chart

Keep the query data cached, so that it can be re-filtered without a query.

Addationally, fix up the logic to get/apply data filters