Scatter plot: Fixes for not resizing in Firefox and general cleanup
The scatter plot was failing to resize in Firefox because we were recreating the Chart object, and the resize
monitoring div was lost when the chart was recreated.
Additionally, for there was some clash with naming the child view 'scatter' and having the data for the chart
in a variable also named 'scatter' - so rename the data to 'chart' instead.