Total rewrite of the filter service.
Eliminated the idea of relations (for now) to fix a number of issues
that couldn't be fixed with how the filter service was previously set
up. Changed all visualizations over to use the new filter service
format, and though some of them are a mess inside (the map in
particular: there's a lot of work to be done in figuring out how filters
on layered components should even work) everything present in the demo
works.