Document Viewer SortField removed
Sortfield merely replaced and modified datefield so I removed sortfield in favor of datefield for now, but we do want to integrated sortfield in the future. Sortfield allowed a useful feature to soft it by any field instead of just date. TODO: we want both sortfield and datefield working in the future